Create Tradier base class and Account derived class
This defines source code for the Tradier and Account classes. It also
defines a simple Account method to retrieve basic account information
from Tradier.

The CMakeLists.txt file defines the build process for the code. This
will probably need to be edited in the future though because it is
hard-coded to accomodate the fact that the nlohmann-json header file
was installed on my mac with homebrew. I don't believe this is a common
/flexible practice.

This also defines the EnvReader file. EnvReader acts like the dotenv
python packge. It assumes that:
	• The TradierCPP executable is run from the build directory
	• There exists a .env file in the root directory
	• The .env file contains keys tradier_acct and tradier_token