Integrate QuantLib into TradierCPP
This updates CMakeLists to link and include the QuantLib library during
the build process. It also creates source code for a new executable,
TradierQLib, that confirms QuantLib is successfully included. The next
step is to use equity/options data from Tradier as parameters for a
Black Scholes options pricing model and implementing buy/sell logic.